% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_Main.R
\name{getDataAnnotationByEnsemblGeneID}
\alias{getDataAnnotationByEnsemblGeneID}
\title{Get expression data by ensembl gene id}
\usage{
getDataAnnotationByEnsemblGeneID(myEnsemblGeneIDs = NULL, printURL = FALSE,
  printTime = FALSE, myStudy = NULL, myNorms = NULL)
}
\arguments{
\item{myEnsemblGeneIDs}{character vector of ensembl gene id to query. Use \code{\link{getEnsemblGeneIDs}} to get all available ensembl gene ids.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}

\item{myStudy}{character vector of studies to query. If NULL, returns data for all studies. Use \code{\link{getStudies}} to get all available studies.}

\item{myNorms}{character vector of normalizations to query. If NULL, returns data for all normalizations. Current normalizations include 'rsem' (gene level), 'sample_abundance'(transcript level), and 'sample_rsem_isoform'(transcript level).}
}
\value{
dataframe of sample and expression data corresponding to the query.
}
\description{
Get expression data by ensembl gene id (Gencode ID)
}
\examples{
getDataAnnotationByEnsemblGeneID(myEnsemblGeneIDs = c("ENSG00000002079.12", "ENSG00000134323.10"),
  printURL = TRUE, printTime = TRUE, myStudy = c("TARGET", "TCGA"),
  myNorms = NULL)
getDataAnnotationByEnsemblGeneID(myEnsemblGeneIDs = c("ENSG00000002079.12", "ENSG00000134323.10"),
  printURL = TRUE, printTime = TRUE, myStudy = "GTEx",
  myNorms = c("rsem", "sample_rsem_isoform"))
getDataAnnotationByEnsemblGeneID(myEnsemblGeneIDs = c("ENSG00000002079.12", "ENSG00000134323.10"),
  printURL = TRUE, printTime = TRUE, myStudy = NULL,
  myNorms = NULL)

}
