% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_Main.R
\name{getDataAnnotationByGeneSymbol}
\alias{getDataAnnotationByGeneSymbol}
\title{Get expression data by hugo gene symbol}
\usage{
getDataAnnotationByGeneSymbol(
  myGeneSymbols = NULL,
  printURL = FALSE,
  printTime = FALSE,
  myStudy = NULL,
  myNorms = NULL
)
}
\arguments{
\item{myGeneSymbols}{character vector of hugo gene symbol to query. Use \code{\link{getGeneSymbols}} to get all available gene symbols.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}

\item{myStudy}{character vector of studies to query. If NULL, returns data for all studies. Use \code{\link{getStudies}} to get all available studies.}

\item{myNorms}{character vector of normalizations to query. If NULL, returns data for all normalizations. Current normalizations include 'rsem' (gene level), 'sample_abundance'(transcript level), and 'sample_rsem_isoform'(transcript level).}
}
\value{
dataframe of sample and expression data corresponding to the query.
}
\description{
Get expression data by hugo gene symbol
}
\examples{
getDataAnnotationByGeneSymbol(myGeneSymbols = c("MYCN","MYC"),
  printURL = TRUE, printTime = TRUE, myStudy = c("TARGET", "TCGA"),
  myNorms = NULL)
getDataAnnotationByGeneSymbol(myGeneSymbols = c("MYCN","MYC"),
  printURL = TRUE, printTime = TRUE, myStudy = "GTEx",
  myNorms = c("rsem", "sample_rsem_isoform"))
getDataAnnotationByGeneSymbol(myGeneSymbols = c("MYCN","MYC"),
  printURL = TRUE, printTime = TRUE, myStudy = NULL,
  myNorms = NULL)

}
