% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getEnsemblGeneIDs}
\alias{getEnsemblGeneIDs}
\title{Get available ensembl gene ids}
\usage{
getEnsemblGeneIDs(printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
string of available ensembl gene ids
}
\description{
Get available ensembl gene ids
}
\examples{
getEnsemblGeneIDs()

}
