% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getEnsemblGeneIDsInfo}
\alias{getEnsemblGeneIDsInfo}
\title{Get feature info by ensembl gene id}
\usage{
getEnsemblGeneIDsInfo(myEnsemblGeneIDs = NULL, printURL = FALSE,
  printTime = FALSE)
}
\arguments{
\item{myEnsemblGeneIDs}{character vector of ensembl ids to query. Use \code{\link{getEnsemblGeneIDs}} to get all available ensembl gene ids.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
dataframe of features corresponding to query ids
}
\description{
Get feature info by ensembl gene id
}
\examples{
getEnsemblGeneIDsInfo(myEnsemblGeneIDs = "ENSG00000002079.12")
getEnsemblGeneIDsInfo(myEnsemblGeneIDs = c("ENSG00000002079.12", "ENSG00000134323.10"))

}
