% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getGeneDataFromGeneSymbol}
\alias{getGeneDataFromGeneSymbol}
\title{Get expression data by hugo gene symbol}
\usage{
getGeneDataFromGeneSymbol(
  myGeneSymbols = NULL,
  myProjection = "summary",
  printURL = FALSE,
  printTime = FALSE,
  myStudy = NULL,
  myNorms = NULL
)
}
\arguments{
\item{myGeneSymbols}{character vector of hugo gene symbols to query. Use \code{\link{getGeneSymbols}} to get all available gene symbols.}

\item{myProjection}{character string. Allowed values: 'detailed' or 'summary'(by default).}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}

\item{myStudy}{character vector of studies to query. If NULL, returns data for all studies. Use \code{\link{getStudies}} to get all available studies.}

\item{myNorms}{character vector of normalizations to query. If NULL, returns data for all normalizations. Current normalizations include 'rsem' (gene level), 'sample_abundance'(transcript level), and 'sample_rsem_isoform'(transcript level).}
}
\value{
dataframe of expression data corresponding to query symbols
}
\description{
Get expression data by hugo gene symbol
}
\examples{
getGeneDataFromGeneSymbol(myGeneSymbols = c("MYCN", "MYC"), printURL = TRUE, printTime = TRUE, myStudy = c("TARGET", "TCGA"), myNorms = NULL)
getGeneDataFromGeneSymbol(myGeneSymbols = c("MYCN", "MYC"), printURL = TRUE, printTime = TRUE, myStudy = "GTEx", myNorms = c("rsem", "sample_rsem_isoform"))
getGeneDataFromGeneSymbol(myGeneSymbols = c("MYCN", "MYC"), printURL = TRUE, printTime = TRUE, myStudy = NULL, myNorms = NULL)

}
