% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getGeneSymbols}
\alias{getGeneSymbols}
\title{Get available hugo gene symbols}
\usage{
getGeneSymbols(printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
string of available hugo gene symbols
}
\description{
Get available hugo gene symbols
}
\examples{
getGeneSymbols()

}
