% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getGeneSymbolsInfo}
\alias{getGeneSymbolsInfo}
\title{Get feature info by hugo gene symbol}
\usage{
getGeneSymbolsInfo(myGeneSymbols = NULL, printURL = FALSE,
  printTime = FALSE)
}
\arguments{
\item{myGeneSymbols}{character vector of hugo gene symbols to query. Use \code{\link{getGeneSymbols}} to get all available ensembl gene symbols.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
dataframe of features corresponding to query symbols
}
\description{
Get feature info by hugo gene symbol
}
\examples{
getGeneSymbolsInfo(myGeneSymbols = "MYCN")
getGeneSymbolsInfo(myGeneSymbols = c("MYCN", "MYC"))

}
