% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getSamples}
\alias{getSamples}
\title{Get sample metadata}
\usage{
getSamples(
  myStudy = NULL,
  allStudies = TRUE,
  printURL = FALSE,
  printTime = FALSE
)
}
\arguments{
\item{myStudy}{character vector of one or more studies. Use \code{\link{getStudies}} to get all available studies.}

\item{allStudies}{TRUE by default. If one or more studies are specified, it is set to FALSE.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
dataframe of sample metadata
}
\description{
Get sample metadata
}
\examples{
getSamples()
getSamples(myStudy = c("TCGA","TARGET"))

}
