% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getStudies}
\alias{getStudies}
\title{Get all studies}
\usage{
getStudies(printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
string of available studies
}
\description{
Get all studies
}
\examples{
getStudies()

}
