% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getTranscriptInfo}
\alias{getTranscriptInfo}
\title{Get feature info by ensembl transcript id}
\usage{
getTranscriptInfo(myTranscripts = NULL, printURL = FALSE,
  printTime = FALSE)
}
\arguments{
\item{myTranscripts}{character vector of ensembl transcript ids to query. Use \code{\link{getTranscripts}} to get all available ensembl transcripts ids.}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
dataframe of features corresponding to query ids
}
\description{
Get feature info by ensembl transcript id
}
\examples{
getTranscriptInfo(myTranscripts = "ENST00000488147.1")
getTranscriptInfo(myTranscripts = c("ENST00000488147.1","ENST00000425880.1"))

}
