% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{getTranscripts}
\alias{getTranscripts}
\title{Get available ensembl transcript ids}
\usage{
getTranscripts(printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
string of available ensembl transcript ids
}
\description{
Get available ensembl transcript ids
}
\examples{
getTranscripts()

}
