% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{retrieveDataJSON}
\alias{retrieveDataJSON}
\title{Retrieve data as JSON object}
\usage{
retrieveDataJSON(myURL = NULL, printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{myURL}{character string representing URL}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
json output
}
\description{
Retrieve data as JSON object
}
\examples{
api <- returnBaseEndPoint()
retrieveDataJSON(myURL = paste0(api, "data/transcripts/ids/ENST00000488147.1/studies/TARGET/normalizations/rsem?projection=summary"))

}
