% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{retrieveDataTAB}
\alias{retrieveDataTAB}
\title{Retrieve data in TSV format}
\usage{
retrieveDataTAB(myURL = NULL, printURL = FALSE, printTime = FALSE)
}
\arguments{
\item{myURL}{character string representing URL}

\item{printURL}{logical to print URL}

\item{printTime}{logical to print time taken}
}
\value{
dataframe
}
\description{
Retrieves data in TSV format
}
\examples{
api <- returnBaseEndPoint()
retrieveDataTAB(myURL = paste0(api,"data/transcripts/ids/ENST00000488147.1/studies/TARGET/normalizations/rsem?projection=summary"))

}
