% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDE_AggMergeVis.R
\name{returnBaseEndPoint}
\alias{returnBaseEndPoint}
\title{Get API base URL}
\usage{
returnBaseEndPoint()
}
\value{
character string representing the base url.
}
\description{
Returns api base url
}
\examples{
returnBaseEndPoint()

}
